pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "englog"
version = "0.2.0"
description = "Automatic engineering logbook — capture decisions, not just tasks."
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "mss>=9.0",
    "Pillow>=10.0",
    "psutil>=5.9",
    "pystray>=0.19",
    "requests>=2.31",
    "openpyxl>=3.1",
    "customtkinter>=5.2",
    "keyboard>=0.13",
    "pywin32>=306; sys_platform == 'win32'",
]

[project.scripts]
englog = "englog.cli:cli"
englog-gui = "englog.gui:main"

[tool.setuptools.packages.find]
include = ["englog*"]
